Rework CROWN tasks. Reduce number of tasks and files. Use multiple quantities_map files for compilation instead of fused file.
