+++ lawluigi_configs/KingMaker_law.cfg
@@ -1,18 +1,8 @@
 [modules]
-BuildCROWNLib
-ConfigureDatasets
 CROWNBase
-CROWNBuild
-CROWNBuildFriend
-CROWNBuildMultiFriend
-CROWNFriends
-CROWNMultiFriends
-CROWNRun
-FriendQuantitiesMap
-ProduceFriends
-ProduceMultiFriends
-ProduceSamples
-QuantitiesMap
+CROWNMain
+CROWNFriend
+ProduceNtuples
 
 [logging]
 # law: DEBUG
