+++ lawluigi_configs/KingMaker_luigi.cfg
@@ -80,8 +80,6 @@ force_repack_tarball = False
 
 [CROWNBuildFriend]
 
-[CROWNBuildMultiFriend]
-
 [BuildCROWNLib]
 
 [CROWNRun]
@@ -92,35 +90,18 @@ htcondor_request_disk = 20000000
 ; for these eras, only one file per task is processed
 problematic_eras = ["2018B", "2017C", "2016B-ver2"]
 
-[CROWNFriends]
+[CROWNFriend]
 ; HTCondor
 htcondor_walltime = 10800
 htcondor_request_memory = 16000
 htcondor_request_disk = 20000000
 ; friends have to be run in single core mode to ensure a correct order of the tree entries
 htcondor_request_cpus = 1
-
-[CROWNMultiFriends]
-; HTCondor
-htcondor_walltime = 10800
-htcondor_request_memory = 16000
-htcondor_request_disk = 20000000
-; friends have to be run in single core mode to ensure a correct order of the tree entries
-htcondor_request_cpus = 1
-
-[ProduceFriends]
-
-[ProduceMultiFriends]
-; the mapping is "config_name": "friend_name"
-friend_mapping = {}
-; friend_mapping = {
-;     "unittest_friends": "id_iso_weights",
-;     "unittest_friends_2": "svfit"}
-
-[ProduceSamples]
 
 [ConfigureDatasets]
 silent = True
 ; set to False to print out the datasets
 
 [QuantitiesMap]
+
+[ProduceNtuples]
